requirements.txt
youtube-transcript-api
langchain-community
langchain
faiss-cpu 
tiktoken 
python-dotenv
torch
numpy